+++ Cappuccino.DAL/SystemManage/SysDataAuthorizeDao.cs
@@ -8,7 +8,7 @@ namespace Cappuccino.DAL
     public class SysDataAuthorizeDao : BaseDao<SysDataAuthorizeEntity>, ISysDataAuthorizeDao
     {
         // 根据角色ID删除数据权限（AuthorizeId=角色ID）
-        public void DeleteByRoleId(int roleId)
+        public void DeleteByRoleId(long roleId)
         {
             var entities = GetList(x => x.AuthorizeId == roleId);
