+++ Cappuccino.Entity/SystemManage/SysNoticeEntity.cs
@@ -11,7 +11,7 @@ namespace Cappuccino.Entity
 
         public int NoticeType { get; set; }
 
-        public int NoticeSender { get; set; }
+        public long NoticeSender { get; set; }
 
         public int NoticeAccept { get; set; }
