+++ Cappuccino.Entity/SystemManage/SysUserEntity.cs
@@ -7,7 +7,7 @@ namespace Cappuccino.Entity
     {
         public string UserName { get; set; }
         public string NickName { get; set; }
-        public int? DepartmentId { get; set; }
+        public long? DepartmentId { get; set; }
         public string PasswordHash { get; set; }
         public string PasswordSalt { get; set; }
         public string HeadIcon { get; set; }
