+++ Cappuccino.IBLL/SystemManage/ISysAutoJobService.cs
@@ -10,20 +10,20 @@ namespace Cappuccino.IBLL.System
         /// </summary>
         /// <param name="id">任务ID</param>
         /// <returns>是否启动成功</returns>
-        Task<bool> StartJob(int id);
+        Task<bool> StartJob(long id);
 
         /// <summary>
         /// 停止任务
         /// </summary>
         /// <param name="id">任务ID</param>
         /// <returns>是否停止成功</returns>
-        Task<bool> StopJob(int id);
+        Task<bool> StopJob(long id);
 
         /// <summary>
         /// 立即执行任务
         /// </summary>
         /// <param name="id">任务ID</param>
         /// <returns>是否执行成功</returns>
-        Task<bool> ExecuteJob(int id);
+        Task<bool> ExecuteJob(long id);
     }
 }
