+++ Cappuccino.Model/SystemManage/UserActionModel.cs
@@ -2,8 +2,8 @@
 {
     public class UserActionModel
     {
-        public int Id { get; set; }
-        public int ParentId { get; set; }
+        public long Id { get; set; }
+        public long ParentId { get; set; }
         public string Name { get; set; }
         public string Code { get; set; }
